+++ src/apps/json-viewer/components/JsonInput.svelte
@@ -11,11 +11,13 @@
     }
 </script>
 
+<span class="label">JSON:</span>
 <textarea class="json-input" bind:value={jsonInput} on:change={updateJsonStore}></textarea>
 
 <style lang="scss">
     .json-input {
         width: 100%;
         height: 50%;
+        margin-top: .5em;
     }
 </style>