+++ src/apps/whiteboard/components/DrawArea.svelte
@@ -117,7 +117,6 @@
     #canvas {
         width: 100%;
         height: 100%;
-        border: 2px solid #ccc;
         border-radius: 5px;
     }
 </style>