+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 Changelog
 ---
 
+v1.2.0
+--
+New Apps:
+- JWT decoder app
+- Base64 encode/decode app
+
+Other:
+- Minor style adjustments for buttons
+- Global font change to monospace
+
+v1.1.0
+--
+- Updated app icon
+- Dark mode re-design
+
 v1.0.0
 --
 Initial release containing 4 apps:
@@ -8,8 +23,3 @@ Initial release containing 4 apps:
 - notepad
 - sha hasher
 - whiteboard
-
-v1.1.0
---
-- Updated app icon
-- Dark mode re-design